+++ server/src/main/java/org/apache/iotdb/db/engine/merge/task/RecoverMergeTask.java
@@ -20,8 +20,8 @@
 package org.apache.iotdb.db.engine.merge.task;
 
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.db.engine.merge.recover.LogAnalyzer;
-import org.apache.iotdb.db.engine.merge.recover.LogAnalyzer.Status;
+import org.apache.iotdb.db.engine.merge.recover.MergeLogAnalyzer;
+import org.apache.iotdb.db.engine.merge.recover.MergeLogAnalyzer.Status;
 import org.apache.iotdb.db.engine.merge.recover.MergeLogger;
 import org.apache.iotdb.db.engine.merge.selector.MaxSeriesMergeFileSelector;
 import org.apache.iotdb.db.engine.storagegroup.TsFileResource;
@@ -31,11 +31,12 @@ import org.apache.iotdb.db.utils.MergeUtils;
 import org.apache.iotdb.tsfile.file.metadata.ChunkMetadata;
 import org.apache.iotdb.tsfile.write.writer.RestorableTsFileIOWriter;
 
+import org.h2.store.fs.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
 import java.util.ArrayList;
@@ -51,7 +52,7 @@ public class RecoverMergeTask extends MergeTask {
 
   private static final Logger logger = LoggerFactory.getLogger(RecoverMergeTask.class);
 
-  private LogAnalyzer analyzer;
+  private MergeLogAnalyzer analyzer;
 
   public RecoverMergeTask(
       List<TsFileResource> seqFiles,
@@ -73,7 +74,7 @@ public class RecoverMergeTask extends MergeTask {
     }
     long startTime = System.currentTimeMillis();
 
-    analyzer = new LogAnalyzer(resource, taskName, logFile, storageGroupName);
+    analyzer = new MergeLogAnalyzer(resource, taskName, logFile, storageGroupName);
     Status status = analyzer.analyze();
     if (logger.isInfoEnabled()) {
       logger.info(
@@ -104,7 +105,7 @@ public class RecoverMergeTask extends MergeTask {
     }
   }
 
-  private void resumeAfterFilesLogged(boolean continueMerge) throws IOException {
+  private void resumeAfterFilesLogged(boolean continueMerge) throws IOException, MetadataException {
     if (continueMerge) {
       resumeMergeProgress();
       calculateConcurrentSeriesNum();
@@ -214,7 +215,7 @@ public class RecoverMergeTask extends MergeTask {
           tsFileResource.getTsFile().getName(),
           fileCnt,
           resource.getSeqFiles().size());
-      RestorableTsFileIOWriter mergeFileWriter = resource.getMergeFileWriter(tsFileResource);
+      RestorableTsFileIOWriter mergeFileWriter = resource.getMergeFileWriter(tsFileResource, true);
       mergeFileWriter.makeMetadataVisible();
       mergeContext.getUnmergedChunkStartTimes().put(tsFileResource, new HashMap<>());
       List<PartialPath> pathsToRecover = analyzer.getMergedPaths();
@@ -287,12 +288,15 @@ public class RecoverMergeTask extends MergeTask {
     for (Entry<File, Long> entry : analyzer.getFileLastPositions().entrySet()) {
       File file = entry.getKey();
       Long lastPosition = entry.getValue();
-      if (file.exists() && file.length() != lastPosition) {
-        try (FileInputStream fileInputStream = new FileInputStream(file)) {
-          FileChannel channel = fileInputStream.getChannel();
+      if (file != null && file.exists() && file.length() != lastPosition && lastPosition != 0) {
+        try (FileOutputStream fileOutputStream = new FileOutputStream(file, true)) {
+          FileChannel channel = fileOutputStream.getChannel();
           channel.truncate(lastPosition);
           channel.close();
         }
+      } else if (file != null && lastPosition == 0) {
+        FileUtils.delete(file.getPath());
+        resource.closeAndRemoveWriter(file);
       }
     }
     analyzer.setFileLastPositions(null);